commit mp3 renderer

+++ typings/lame.d.ts
@@ -1,7 +1,7 @@
 declare module 'lamejs'{
     export class Mp3Encoder {
         constructor(channels: number, sampleRate: number, bitRate: number);
-        encodeBuffer(left: Float32Array, right?: Float32Array): Int8Array;
+        encodeBuffer(left: Int16Array, right?: Int16Array): Int8Array;
         flush(): Int8Array;
     }
